[update] enforce float32 and fix gym version to 0.18.0
* The gym with higher version will check whether observation_space contains obs.
However, pybullet joint limit are satisfied by soft physical dynamics,
which will lead to joint-position exceeding joint-limit.

+++ scibotpark/locomotion/base.py
@@ -71,7 +71,7 @@ class LocomotionEnv(PybulletEnv, Env):
                 inertial_data["linear_velocity"],
                 inertial_data["rotation"],
                 inertial_data["angular_velocity"],
-            ])
+            ]).astype(np.float32)
         if len(self.obs_types) == 1:
             return obs[self.obs_types[0]]
         else:
